Change evidence note compliance year to be int to match datetime.year

+++ src/EA.Weee.Core/AatfEvidence/EvidenceNoteData.cs
@@ -39,7 +39,7 @@
 
         public string RejectedReason { get; set; }
 
-        public short ComplianceYear { get; set; }
+        public int ComplianceYear { get; set; }
 
         public List<EvidenceTonnageData> EvidenceTonnageData { get; set; }
